% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auth_token.R
\name{get_auth_token}
\alias{get_auth_token}
\title{Get auth token to query Dynamodb Data from JsPsych Production Server ----}
\usage{
get_auth_token(api = NA)
}
\description{
Get auth token to query Dynamodb Data from JsPsych Production Server ----
}
