% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unnest_s3.R
\name{get_unnest_s3}
\alias{get_unnest_s3}
\title{Query Dynamodb Data from JsPsych Production Server, return S3 file as JSON}
\usage{
get_unnest_s3(api = NA, token = NA, experiment = NA, fields = NA)
}
\description{
Query Dynamodb Data from JsPsych Production Server, return S3 file as JSON
}
