% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_study_ddb.R
\name{query_study_ddb}
\alias{query_study_ddb}
\title{Query Dynamodb Data from JsPsych Production Server ----}
\usage{
query_study_ddb(db = "dynamodb", api = NA, study = NA, token = NA, fields = NA)
}
\description{
Query Dynamodb Data from JsPsych Production Server ----
}
